% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthapis.R
\docType{data}
\name{abm_units}
\alias{abm_units}
\title{Public data for KTH organizational units used in the Annual Bibliometric Monitoring.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 35 rows and 19 columns.}
\usage{
abm_units
}
\description{
A table combining data from the \code{bibliomatrix} package with the KTH Directory API and
the Altmetric Exlorer dataset.
Departments at KTH used in the ABM application are listed with their
DiVA Unit identifiers, kthid identifiers and Altmetric identifiers.
}
\keyword{datasets}
