% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthapis.R
\docType{data}
\name{altmetric_explorer}
\alias{altmetric_explorer}
\title{Public data with KTH Departments at Altmetric}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 590 rows and 6 columns.
}
\source{
\url{https://www.altmetric.com/explorer/institution/departments}
}
\usage{
altmetric_explorer
}
\description{
Departments at KTH are listed with their DiVA Unit identifiers and English names.
}
\keyword{datasets}
