% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthapis.R
\docType{data}
\name{altmetric_explorer_attention}
\alias{altmetric_explorer_attention}
\title{Public data with KTH Altmetric Outputs with Attention}
\format{
An object of class \code{numeric} of length 1.
}
\source{
\url{https://www.altmetric.com/explorer/api/research_outputs/attention}
}
\usage{
altmetric_explorer_attention
}
\description{
Data fetched from Altmetric Explorer API
}
\keyword{datasets}
