% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_swepub}
\alias{classify_swepub}
\title{Classify into subject categories using SwePub Web API}
\usage{
classify_swepub(title, abstract, keywords, level = 3)
}
\arguments{
\item{title}{the title of the paper}

\item{abstract}{the abstract}

\item{keywords}{keywords used}

\item{level}{either 3 or 5, default: 3}
}
\description{
Use the title, abstract and keywords to get a suggestion for subject
categories to use from the Swedish standard set of 3 or 5 digit
subject categories.
}
\details{
Calls the API at https://swepub-qa.kb.se/api/v1/classify
}
\examples{
\dontrun{
classify_swepub(
   title = "Magnetic resonance",
   abstract = "This paper deals with magnetic resonance",
   keywords = "magnetic radiotherapy nuclear",
   level = 5
)
}
}
