% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{config}
\alias{config}
\title{KTH APIs configuration}
\usage{
config()
}
\description{
This function provides a list with settings for the base url and user agent
used in API calls, and the relevant API keys required.
}
\details{
Please edit your ~./Renviron file and set values for environment variables
KTH_API_KEY_DIRECTORY, KTH_API_KEY_PROFILES and KTH_API_KEY_PUBLICATIONS
you can use file.edit("~/.Renviron"); readRenviron("~/.Renviron"))
}
