% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{issues_swepub}
\alias{issues_swepub}
\title{DiVA publication records from KTH with subject classification issues}
\usage{
issues_swepub(
  institution = "kth",
  year_beg = 2012L,
  year_end = 2020L,
  page_size = 20
)
}
\arguments{
\item{institution}{the institution string, Default: 'kth'}

\item{year_beg}{filter from year, Default: 2012}

\item{year_end}{filter to year, Default: 2020}

\item{page_size}{for paged responses, the page length, Default: 20}
}
\value{
a tibble with results
}
\description{
DiVA publication records from KTH with subject classification issues
}
\details{
Data comes from \href{https://bibliometri.swepub.kb.se/process}{}
}
\examples{
\dontrun{
if(interactive()){
 issues_swepub()
 }
}
}
