% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_belonging_institutional}
\alias{kth_belonging_institutional}
\title{Retrieve organizational belonging for a given kthid or accountname at
institutional level (assumed to be the second highest hierarchical level)}
\usage{
kth_belonging_institutional(kthid, cfg = config())
}
\arguments{
\item{kthid}{a string with the account name or KTH user id}

\item{cfg}{configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
a tibble with the userid, the "slug" and the org unit description
}
\description{
Retrieve organizational belonging for a given kthid or accountname at
institutional level (assumed to be the second highest hierarchical level)
}
\examples{
\dontrun{
kth_belonging_institutional("u1z88syr")
kth_belonging_institutional("tjep")
}
}
