% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{kth_catalog}
\alias{kth_catalog}
\title{Retrieve catalog data from the KTH Directory API given a slug for an organizational unit}
\usage{
kth_catalog(cfg = NULL, slug = NULL, lang = NULL)
}
\arguments{
\item{cfg}{a configuration setting for the KTH APIs including base URL etc, by default from config()}

\item{slug}{the slug (a kind of human readable organizational identifer used at KTH) to enumerate}

\item{lang}{language, by default "en" is used, can also be set to "sv"}
}
\value{
results records returned from the search
}
\description{
Retrieve catalog data from the KTH Directory API given a slug for an organizational unit
}
\examples{
\dontrun{
kth_catalog("s")
}
}
