% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva.R
\name{kth_diva_aliases}
\alias{kth_diva_aliases}
\title{Name aliases for DiVA authors in the KTH DiVA portal}
\usage{
kth_diva_aliases(authors = kth_diva_authors())
}
\arguments{
\item{authors}{a tibble with authors data, default: kth_diva_authors()}
}
\value{
data frame with results
}
\description{
This function returns data for authors that for the same author identifier
have multiple names registered in composite DiVA Name strings
(bibliographic names for a publication).
}
