% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva_pubs.R
\name{kth_diva_pubs}
\alias{kth_diva_pubs}
\title{Retrieve DiVA publications for KTH from the KTH DiVA portal}
\usage{
kth_diva_pubs(orgid = "177", year_beg = "2012", year_end = "2019")
}
\arguments{
\item{orgid}{the DiVA organisation id, by default "177" for KTH}

\item{year_beg}{the beginning of the period, by default "2012"}

\item{year_end}{the end of the period, by default "2019"}
}
\value{
data frame with results
}
\description{
This function sends a request to KTH's DiVA portal for a CSV data export
covering KTH publications between 2012 and 2019.
}
