% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_profile}
\alias{kth_profile}
\title{Retrieve data for KTH Profiles}
\usage{
kth_profile(
  kthid = NULL,
  orcid = NULL,
  socialid = NULL,
  username = NULL,
  config = NULL
)
}
\arguments{
\item{kthid}{the kthId for the user profile}

\item{orcid}{the ORC identifier for the user profile}

\item{socialid}{the social id for the user profile}

\item{username}{the accountname for the user profile}

\item{config}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
See details at \url{https://api.kth.se/api/profile/swagger/index.html}
}
\examples{
\dontrun{
kth_profile(username = "agnel")
kth_profile(username = "markussk")
tryCatch(kth_profile(username = "hoyce"), error = function(e) e, warning("Does not exist?"))
}
}
