% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_profile_legacy}
\alias{kth_profile_legacy}
\title{Retrieve data for KTH Profiles}
\usage{
kth_profile_legacy(userid = NULL, config = NULL)
}
\arguments{
\item{userid}{a string with the account name or KTH user id}

\item{config}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
This fcn uses the authenticated API endpoint to mimic a previously available legacy endpoint.
}
\examples{
\dontrun{
kth_profile_legacy("agnel")
kth_profile_legacy("tjep")
tryCatch(kth_profile_legacy("hoyce"), error = function(e) e, warning("Does not exist?"))
}
}
