% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_profile_school_dep}
\alias{kth_profile_school_dep}
\title{Retrieve organizational unit for a given kthid or accountname}
\usage{
kth_profile_school_dep(kthid, cfg = config())
}
\arguments{
\item{kthid}{a string with the account name or KTH user id}

\item{cfg}{configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\description{
Retrieve organizational unit for a given kthid or accountname
}
\examples{
\dontrun{
kth_profile_school_dep("u1z88syr")
kth_profile_legacy("hoyce")
}
}
