% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projects.R
\name{kth_projects}
\alias{kth_projects}
\title{Retrieve data for KTH Projects}
\usage{
kth_projects(
  path = "projects/",
  year_beg = NULL,
  year_end = NULL,
  kthUserName = NULL,
  orcid = NULL,
  tag = NULL,
  config = NULL
)
}
\arguments{
\item{path}{string to indicate endpoint to use, for example
"projects", "projects/public", "projects/hidden", "fundings/", Default: "projects/"}

\item{year_beg}{starting year, for example 2012}

\item{year_end}{ending year, for example 2019}

\item{kthUserName}{username of KTH profile, for example "stemme"}

\item{orcid}{Orcid of members to get, for example "0000-0003-2983-5573"}

\item{tag}{project tag}

\item{config}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
resulting records
}
\description{
See details at \url{https://api-r.referens.sys.kth.se/api/projects/swagger}
}
\examples{
\dontrun{
kth_projects()
kth_projects("projects/public")$content$projects \%>\% tibble::as_tibble()
kth_projects("admin/last-manual-import")
kth_projects("admin/weekly-import-information")
}
}
