% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{kth_publications}
\alias{kth_publications}
\title{Retrieve data for KTH Publications}
\usage{
kth_publications(
  path = c("filteredPublications", "userstatus", "stats", "activePublicUsers",
    "activeNotPublicUsers", "activeUsersWithPublications",
    "activeUsersWithoutPublications", "organisation", "organisations"),
  username = NULL,
  orgid = NULL,
  is_html = NULL,
  l = NULL,
  style = NULL,
  divaUri = NULL,
  q = NULL,
  config = NULL
)
}
\arguments{
\item{path}{one of a set of valid API calls}

\item{username}{the accountname parameter to use}

\item{orgid}{the organisation identifier to use}

\item{is_html}{boolean to indicate if HTML should be returned}

\item{l}{the language code (one of "en" or "sv")}

\item{style}{the style to use (when requesting organisation path)}

\item{divaUri}{the diva URI to use (when requesting organisation/diva path)}

\item{q}{regexp filter (when requesting organizations path)}

\item{config}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
See details at \url{https://api.kth.se/api/publications/swagger/?url=/api/publications/swagger.json#/}
}
\examples{
\dontrun{
kth_publications(username = "hoyce")
}
}
