% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{kth_school_dep}
\alias{kth_school_dep}
\title{Retrieve units in the KTH Directory for the first two levels (school, dep)}
\usage{
kth_school_dep(cfg)
}
\arguments{
\item{cfg}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
Retrieve units in the KTH Directory for the first two levels (school, dep)
}
\examples{
\dontrun{
kth_school_dep()
}
}
