% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi-package.R
\docType{package}
\name{kthapi-package}
\alias{kthapi-package}
\title{kthapi: Data from KTH APIs}
\description{
The KTH APIs provides information about employee profiles, published web content, places, course schemas and program catalogues. This R package interfaces with the API, making data available to use from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KTH-Library/kthapi}
  \item \url{https://kth-library.github.io/kthapi/}
  \item Report bugs at \url{https://github.com/KTH-Library/kthapi/issues}
}

}
\author{
\strong{Maintainer}: Markus Skyttner \email{markussk@kth.se}

Authors:
\itemize{
  \item Gaël Dubus \email{dubus@kth.se}
  \item Agne Larsson \email{agnel@kth.se}
}

}
\keyword{internal}
