% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kthapi-package.R
\name{kthapi}
\alias{kthapi}
\title{kthapi}
\description{
KTH APIs
}
\details{
KTH offers a number of APIs that provides information about employee
profiles, published web content, places, course schemas and program
catalogues. This R package interfaces with the API, making data
available to use from R.

An introduction to the KTH APIs in Swedish is available here: \url{https://www.kth.se/api/anvand-data-fran-kth-1.57059}

An overview of the KTH APIs can be found here: \url{https://www.kth.se/api}

Individual APIs are documented here:
\itemize{
\item Course schema API: \url{https://www.kth.se/api/schema/v2/apidoc.html}
\item Program catalogue API: \url{https://api.kth.se/api/kopps/v2/}
\item Profiles API (personal contact info): \url{https://api.kth.se/api/profile/swagger/}
\item Directory API : \url{https://api.kth.se/api/directory/swagger/}
\item Places API: \url{https://api.kth.se/api/places/swagger/}
\item Publications API: \url{https://api.kth.se/api/publications/swagger/}
\item Projects API: \url{https://api-r.referens.sys.kth.se/api/projects/swagger}
}
}
