% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ldap_search}
\alias{ldap_search}
\title{Search Active Directory at KTH}
\usage{
ldap_search(ldap_query, ldap_attributes = NULL, cfg = ldap_config())
}
\arguments{
\item{ldap_query}{the LDAP query to issue, such as 'ugKthid=*'}

\item{ldap_attributes}{set of attributes to return, by default NULL
but can be a character vector of attributes, such as c('ugKthid', 'ugOrcid')}

\item{cfg}{the connection credentials, by default given from ldap_config()}
}
\value{
tibble with results
}
\description{
This function uses ldapsearch to query the KTH Active Directory. It requires
environment variables to be set in .Renviron, specifically
LDAP_USER, LDAP_PASS, LDAP_HOST and LDAP_BASE for the service
account used for the queries.
}
\details{
The ldapsearch command makes use of -E pr=2147483647/noprompt to avoid paging
and returns results in LDIF format which is parsed into a tibble.
}
\examples{
\dontrun{
if(interactive()){
 ldap_search("(&(ugOrcid=*)(ugKthid=*))", c("ugKthid", "ugOrcid"))
 }
}
}
