% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva.R
\name{parse_diva_name}
\alias{parse_diva_name}
\title{Parse a DiVA namestring with bibliographic data}
\usage{
parse_diva_name(diva_name)
}
\arguments{
\item{diva_name}{the string with author names and identifiers}
}
\value{
data frame with results
}
\description{
This function parses a string and returns the components (such as
name, kthid and ORCID, org affiliation)
}
