% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diva.R
\name{parse_diva_names}
\alias{parse_diva_names}
\title{Parse all DiVA namestrings in DiVA publications}
\usage{
parse_diva_names(pubs = kth_diva_pubs())
}
\arguments{
\item{pubs}{a data frame with diva publications, by default kth_diva_pubs()}
}
\value{
data frame with results
}
\description{
This function parses all DiVA namestrings in a set of DiVA publications.
}
