% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ug_orcid_kthid_unit}
\alias{ug_orcid_kthid_unit}
\title{Run ldap search for kthid/orcid pairs}
\usage{
ug_orcid_kthid_unit()
}
\value{
list with slots for tibbles
}
\description{
A LDAP search resulting in a list with three tibbles;
for persons with orcid data, persons with unit affiliation and
persons with more than one unit
}
\examples{
\dontrun{
if(interactive()){
 ug_orcid_kthid_unit()
 }
}
}
