% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-kthapis.R
\docType{data}
\name{altmetric_explorer_highlights}
\alias{altmetric_explorer_highlights}
\title{Public data with KTH Altmetric Highlights summary indicators}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\source{
\url{https://www.altmetric.com/explorer/highlights}
}
\usage{
altmetric_explorer_highlights
}
\description{
Public data with KTH Altmetric Highlights summary indicators
}
\keyword{datasets}
