% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{kth_displayname}
\alias{kth_displayname}
\title{Retrieve display name given kthid}
\usage{
kth_displayname(kthid, cfg = config())
}
\arguments{
\item{kthid}{a string with the account name or KTH user id}

\item{cfg}{configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\description{
Retrieve display name given kthid
}
