% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publications.R
\name{kth_publications}
\alias{kth_publications}
\title{Retrieve data for KTH Publications}
\usage{
kth_publications(
  path = c("filteredPublications", "userstatus", "stats", "activePublicUsers",
    "activeNotPublicUsers", "activeUsersWithPublications",
    "activeUsersWithoutPublications", "organisation", "organisations"),
  username = NULL,
  orgid = NULL,
  is_html = c("false", "true"),
  lang = c("sv", "en"),
  style = c("ieee", "apa"),
  divaUri = NULL,
  q = NULL,
  config = NULL
)
}
\arguments{
\item{path}{one of a set of valid API calls}

\item{username}{the accountname parameter to use}

\item{orgid}{the organisation identifier to use}

\item{is_html}{string to indicate if HTML should be returned, default "false"}

\item{lang}{the language code (one of "en" or "sv")}

\item{style}{the style to use (when requesting organisation path)}

\item{divaUri}{the diva URI to use (when requesting organisation/diva path)}

\item{q}{regexp filter (when requesting organizations path)}

\item{config}{a configuration setting for the KTH APIs including base URL etc, by default from config()}
}
\value{
results records returned from the search
}
\description{
See details at \url{https://api.kth.se/api/publications/swagger}
}
\examples{
\dontrun{
kth_publications(path = "userstatus", username = "tjep")

orgid <- tibble::as_tibble(kth_publications(path = "organisations")$content) \%>\%
    dplyr::filter(nameLocalized == "Library") \%>\% dplyr::pull(id)

pubs <- kth_publications("organisation", orgid = orgid)$content$publications \%>\%
    tibble::as_tibble()

uri <- pubs$identifierUri[1]
}
}
